+++ .gitignore
@@ -1,3 +1,3 @@
-.vercel,
-.vscode,
+.vercel
+.vscode
 node_modules
